ml/main.py
import asyncio
import base64
import logging
import os
import sys
from io import BytesIO
from pathlib import Path

# import uvicorn
from config import Config
from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware
from fastapi.staticfiles import StaticFiles
from PIL import Image
from pydantic import BaseModel

