ml/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
fastapi==0.112.2
idna==3.8
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
sniffio==1.3.1
starlette==0.38.4
typing_extensions==4.12.2
